+++ TrainingApp/app/src/main/java/com/traininapp/viewModel/GoalStatCard.java
@@ -12,8 +12,8 @@ import java.util.List;
 public class GoalStatCard implements IStatistic {
 
         String goalName;
-        Integer goalTarget;
-        Integer goalProgress;
+        Double goalTarget;
+        Double goalProgress;
 
     /**
      * The constructor for the class GoalStatCard
@@ -22,7 +22,7 @@ public class GoalStatCard implements IStatistic {
      * @param goalTarget
      * @param goalProgress
      */
-    public GoalStatCard(String goalName, Integer goalTarget, Integer goalProgress) {
+    public GoalStatCard(String goalName, Double goalTarget, Double goalProgress) {
         this.goalName = goalName;
         this.goalTarget = goalTarget;
         this.goalProgress = goalProgress;
@@ -37,11 +37,11 @@ public class GoalStatCard implements IStatistic {
         return goalName;
     }
 
-    public Integer getTarget() {
+    public Double getTarget() {
         return goalTarget;
     }
 
-    public Integer getProgress() {
+    public Double getProgress() {
         return goalProgress;
     }
 }
